Update of kalman_filters
